Renamed as `avoid-accidental-submissions`
